+++ central/policycategory/datastore/datastore_impl_test.go
@@ -27,7 +27,8 @@ type PolicyCategoryDatastoreTestSuite struct {
 	indexer   *indexMocks.MockIndexer
 	datastore DataStore
 
-	ctx context.Context
+	hasReadWritePolicyAccess                 context.Context
+	hasReadWriteWorkflowAdministrationAccess context.Context
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) SetupTest() {
@@ -38,42 +39,52 @@ func (s *PolicyCategoryDatastoreTestSuite) SetupTest() {
 
 	s.datastore = newWithoutDefaults(s.store, s.indexer, nil)
 
-	s.ctx = sac.WithGlobalAccessScopeChecker(context.Background(),
+	s.hasReadWritePolicyAccess = sac.WithGlobalAccessScopeChecker(context.Background(),
 		sac.AllowFixedScopes(
 			sac.AccessModeScopeKeys(storage.Access_READ_ACCESS, storage.Access_READ_WRITE_ACCESS),
 			sac.ResourceScopeKeys(resources.Policy)))
-}
-
-func (s *PolicyCategoryDatastoreTestSuite) TearDownTest() {
-	s.mockCtrl.Finish()
+	// TODO: ROX-13888 Remove duplicated context.
+	s.hasReadWriteWorkflowAdministrationAccess = sac.WithGlobalAccessScopeChecker(context.Background(),
+		sac.AllowFixedScopes(
+			sac.AccessModeScopeKeys(storage.Access_READ_ACCESS, storage.Access_READ_WRITE_ACCESS),
+			sac.ResourceScopeKeys(resources.WorkflowAdministration)))
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) TestAddNewPolicyCategory() {
-	s.store.EXPECT().Upsert(s.ctx, gomock.Any()).Return(nil)
+	s.store.EXPECT().Upsert(s.hasReadWritePolicyAccess, gomock.Any()).Return(nil).Times(1)
 
-	s.indexer.EXPECT().AddPolicyCategory(gomock.Any()).Return(nil)
+	s.indexer.EXPECT().AddPolicyCategory(gomock.Any()).Return(nil).AnyTimes()
 
-	_, err := s.datastore.AddPolicyCategory(s.ctx, fixtures.GetPolicyCategory())
+	_, err := s.datastore.AddPolicyCategory(s.hasReadWritePolicyAccess, fixtures.GetPolicyCategory())
+	s.NoError(err, "expected no error trying to add a category with permissions")
+	// TODO: ROX-13888 Remove duplicated test.
+	s.store.EXPECT().Upsert(s.hasReadWriteWorkflowAdministrationAccess, gomock.Any()).Return(nil).Times(1)
+	_, err = s.datastore.AddPolicyCategory(s.hasReadWriteWorkflowAdministrationAccess, fixtures.GetPolicyCategory())
 	s.NoError(err, "expected no error trying to add a category with permissions")
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) TestAddDuplicatePolicyCategory() {
 	c := fixtures.GetPolicyCategory()
-	s.store.EXPECT().Upsert(s.ctx, c).Return(errors.New("exists"))
+	s.store.EXPECT().Upsert(s.hasReadWritePolicyAccess, c).Return(errors.New("exists"))
 
-	_, err := s.datastore.AddPolicyCategory(s.ctx, c)
+	_, err := s.datastore.AddPolicyCategory(s.hasReadWritePolicyAccess, c)
 	s.Error(err)
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) TestDeletePolicyCategory() {
 	c := fixtures.GetPolicyCategory()
 
-	s.store.EXPECT().Get(s.ctx, "category-id").Return(c, true, nil)
-	s.store.EXPECT().Delete(s.ctx, gomock.Any()).Return(nil).AnyTimes()
+	s.store.EXPECT().Get(s.hasReadWritePolicyAccess, "category-id").Return(c, true, nil).Times(1)
+	s.store.EXPECT().Delete(s.hasReadWritePolicyAccess, gomock.Any()).Return(nil).AnyTimes()
 
 	s.indexer.EXPECT().DeletePolicyCategory(gomock.Any()).Return(nil).AnyTimes()
 
-	err := s.datastore.DeletePolicyCategory(s.ctx, c.Id)
+	err := s.datastore.DeletePolicyCategory(s.hasReadWritePolicyAccess, c.Id)
+	s.NoError(err, "expected no error trying to delete a category with permissions")
+	// TODO: ROX-13888 Remove duplicated test.
+	s.store.EXPECT().Get(s.hasReadWriteWorkflowAdministrationAccess, "category-id").Return(c, true, nil).Times(1)
+	s.store.EXPECT().Delete(s.hasReadWriteWorkflowAdministrationAccess, gomock.Any()).Return(nil).AnyTimes()
+	err = s.datastore.DeletePolicyCategory(s.hasReadWriteWorkflowAdministrationAccess, c.Id)
 	s.NoError(err, "expected no error trying to delete a category with permissions")
 }
 
@@ -81,30 +92,37 @@ func (s *PolicyCategoryDatastoreTestSuite) TestDeleteDefaultPolicyCategory() {
 	c := fixtures.GetPolicyCategory()
 	c.IsDefault = true
 
-	s.store.EXPECT().Get(s.ctx, "category-id").Return(c, true, nil)
+	s.store.EXPECT().Get(s.hasReadWritePolicyAccess, "category-id").Return(c, true, nil)
 
-	err := s.datastore.DeletePolicyCategory(s.ctx, c.Id)
+	err := s.datastore.DeletePolicyCategory(s.hasReadWritePolicyAccess, c.Id)
 	s.Error(err)
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) TestRenamePolicyCategory() {
 	c := fixtures.GetPolicyCategory()
 
-	s.store.EXPECT().Upsert(s.ctx, gomock.Any()).Return(nil).AnyTimes()
-	s.store.EXPECT().Get(s.ctx, c.GetId()).Return(c, true, nil)
+	s.store.EXPECT().Upsert(s.hasReadWritePolicyAccess, gomock.Any()).Return(nil).AnyTimes()
+	s.store.EXPECT().Get(s.hasReadWritePolicyAccess, c.GetId()).Return(c, true, nil)
 
 	s.indexer.EXPECT().AddPolicyCategory(gomock.Any()).Return(nil).AnyTimes()
 
-	c, err := s.datastore.RenamePolicyCategory(s.ctx, c.Id, "Boo's Special Category New Name")
+	c, err := s.datastore.RenamePolicyCategory(s.hasReadWritePolicyAccess, c.Id, "Boo's Special Category New Name")
 	s.NoError(err, "expected no error trying to rename a category with permissions")
 	s.Equal("Boo'S Special Category New Name", c.GetName(), "expected category to be renamed, but it is not")
+
+	// TODO: ROX-13888 Remove duplicated test.
+	s.store.EXPECT().Upsert(s.hasReadWriteWorkflowAdministrationAccess, gomock.Any()).Return(nil).AnyTimes()
+	s.store.EXPECT().Get(s.hasReadWriteWorkflowAdministrationAccess, c.GetId()).Return(c, true, nil)
+	c, err = s.datastore.RenamePolicyCategory(s.hasReadWriteWorkflowAdministrationAccess, c.Id, "Daniel's Special Category New Name")
+	s.NoError(err, "expected no error trying to rename a category with permissions")
+	s.Equal("Daniel'S Special Category New Name", c.GetName(), "expected category to be renamed, but it is not")
 }
 
 func (s *PolicyCategoryDatastoreTestSuite) TestRenamePolicyCategoryDuplicateName() {
-	s.store.EXPECT().Upsert(s.ctx, gomock.Any()).Return(errors.New("exists")).AnyTimes()
-	s.store.EXPECT().Get(s.ctx, "category-id").Return(fixtures.GetPolicyCategory(), true, nil)
+	s.store.EXPECT().Upsert(s.hasReadWritePolicyAccess, gomock.Any()).Return(errors.New("exists")).AnyTimes()
+	s.store.EXPECT().Get(s.hasReadWritePolicyAccess, "category-id").Return(fixtures.GetPolicyCategory(), true, nil)
 
-	_, err := s.datastore.RenamePolicyCategory(s.ctx, "category-id", "new name")
+	_, err := s.datastore.RenamePolicyCategory(s.hasReadWritePolicyAccess, "category-id", "new name")
 	s.Error(err)
 }
 
@@ -112,8 +130,8 @@ func (s *PolicyCategoryDatastoreTestSuite) TestRenameDefaultPolicyCategory() {
 	c := fixtures.GetPolicyCategory()
 	c.IsDefault = true
 
-	s.store.EXPECT().Get(s.ctx, "category-id").Return(c, true, nil)
+	s.store.EXPECT().Get(s.hasReadWritePolicyAccess, "category-id").Return(c, true, nil)
 
-	_, err := s.datastore.RenamePolicyCategory(s.ctx, c.Id, "new name")
+	_, err := s.datastore.RenamePolicyCategory(s.hasReadWritePolicyAccess, c.Id, "new name")
 	s.Error(err)
 }