+++ central/reportconfigurations/datastore/datastore_impl.go
@@ -20,6 +20,7 @@ import (
 )
 
 var (
+	// TODO: ROX-13888 Replace VulnerabilityReports with WorkflowAdministration.
 	reportConfigSAC = sac.ForResource(resources.VulnerabilityReports)
 
 	log = logging.LoggerForModule()